+++ src/components/layouts/auth.tsx
@@ -27,7 +27,7 @@ const authVariants = cva(
         clean: '',
         subtle: 'bg-muted/10',
         brand: 'bg-primary/5',
-        contrast: 'bg-zinc-900'
+        contrast: 'bg-contrast'
       }
     },
     defaultVariants: {